Add comments to document that completion handlers in any `FeedStore` implementation can be invoked in any thread. Clients are responsible to dispatch to appropriate thread, if needed.
