+++ EssentialFeed/EssentialFeed/Feed API/URLSessionHTTPClient.swift
@@ -1,0 +1,35 @@
+import Foundation
+
+// So the production implementation for Client, would be a real request for the API.
+// We need to be able to test it first because we don't have API yet 🤗
+// What will happen at the end of the day is a URLSession -> DataTaskWith(url) request, that return a completion block.
+// So the RemoteClient will have the Session that do that ☝️
+// And will have an API method get(from url) that will be the window to shot for the request.
+// So Client has a Session 😉
+//Test behaviour, not framework (production) details
+// So you free the production from any testing constrains 🆓
+// Implement and maintain Only what you care about 👌
+
+public class URLSessionHTTPClient: HTTPClient {
+
+    private let session: URLSession
+
+    public init(session: URLSession = .shared) {
+        self.session = session
+    }
+
+    private struct UnexpectedValuesRepresentation: Error {}
+
+    public func get(from url: URL, completion: @escaping (ClientResult) -> Void) {
+        session.dataTask(with: url) { data, response, error in
+            if let error = error {
+                completion(.failure(error))
+            }
+            else if let data = data, let response = response as? HTTPURLResponse {
+                completion(.success((data, response)))
+            } else {
+                completion(.failure(UnexpectedValuesRepresentation()))
+            }
+        }.resume()
+    }
+}
