+++ EssentialFeed/EssentialFeed/Feed Cache/FeedStore.swift
@@ -12,7 +12,15 @@ public protocol FeedStore {
     typealias InsertionCompletion = (Error?) -> Void
     typealias RetrievalCompletion = (RetrieveFeedResult) -> Void
 
-    func deleteCachedFeed(completion: @escaping DeletionCompletion)
+    /// The completion handler can be invoked in any thread.
+    /// Clients are responsible to dispatch to appropriate threads, if needed.
     func insert(_ items: [LocalFeedImage], timeStamp: Date, completion: @escaping InsertionCompletion)
+
+    /// The completion handler can be invoked in any thread.
+    /// Clients are responsible to dispatch to appropriate threads, if needed.
     func retrieve(completion: @escaping RetrievalCompletion)
+
+    /// The completion handler can be invoked in any thread.
+    /// Clients are responsible to dispatch to appropriate threads, if needed.
+    func deleteCachedFeed(completion: @escaping DeletionCompletion)
 }
